Nebula
The volume of discarded electronics in East and South- East Asia jumped almost 63% between 2010 and 2015;
however, few efficient recycling treatment is suitable for various E-waste devices. According to exploratory research, LED light stick, a disposables product, is usually utilized for entertainment in a special event. But after the event, it is unlikely that LED light stick would be used again either in the case of remaining at home or thrown into the garbage bin.Our ultimate goal is to reduce the number of disposables electronic devices
that impact our environment. The Nebula is a LED light
stick recycling renting system. Through “To Rent But
Not Buy” proposition, the Nebula allows the LED light
stick to circulate sustainably in the market. At same time, the Nebula proposes special interaction mode such as peer-to-peer which allows the audience to
engage into the atmosphere of concert thoroughly.